Added small script to prepare new machines
This just symlinks the main directory under the "per-machine" directory,
along with a special main file that loads the appropriate files for that
machine. Note this is not the only way to do it; for example, you change
the $NIXOS_CONFIG environment variable to point to that configuration
file. I like this approach as the least intrusive, however.