Added mechanic for randomly switching wallpapers
Note not quite done - it holds the wallpaper path in a global var, which
I would to change, if possible. I also wonder if it is going to eat up
memory if I am not careful. Also requires some renaming of variables and
functions, along with connecting to a singal if possible.