Changed how super+{h,j,k,l,H,J,K,L} worked
Now they have different functions depending on if the client is floating
or not. If the client is tiled/stacked, then those keys will change the
focus and ordering of those windows and only among those non-floating
windows; if the client is floating, then it will reposition and resize
the window, respective from the top left corner of the window. Reording
is done with the [ and ] keys.