Changed zsh prompt
Reversed how the prompt detects if it's in a virtual console or not. Now
it look for a basename with only numbers (Linux and most BSDs) or
starting with the string 'ttyp' (OpenBSD - not sure if this is
universally applied).

Added the tty basename to the prompt for virtual consoles - it's good to
know which console you are using while using them, if only knowing which
buttons to press to change.