Added a bunch of configs
Configs for sc-im, vimpc and youtube-dl. The first two might be changing
their configuations to be XDG conformant, so for now they are symlinked
to the home directories.

Added a shell file for POSIX-compatable sh. Added a file for common
functions.

Added users-dirs.dirs file, to have XDG program recogize my directories
as the various "special" directories (e.g. ~/docs is the Documents
directory).