Modifed zsh config files
* Moved history file to ~/.cache/shell/zsh_history: the history file
fits this location more naturally than the previous location.
TODO: I need add the configuration files for most shell so that they put
their history files in this location and not in $HOME.

* Move aliases and sourcing plugins into seperate files: These files
should be in ~/.config/zsh/. These files are in the dots directory, but
with different versions for each OS and Distro. GNU/Linux distros will
share the same alias files, as they use the same GNU core utils; each of
the BSDs will their own aliases, as the core utils have diverged.

* Added fzf customizations. For A-c (change directory), a preview window
is a directory listing should be shown. C-t shows is using bat only, for
now; in future, it should use an updated peekat script.