Added a bashrc and env files
The bashrc file is intended to very minimal, only to be used on
occation, with only a prompt to show who and where I am and moves the
history to a .cache folder.

The env file is a shell script that holds a bunch of enviroment
variables; most related to default applications and the internal
structure of the $HOME directory.